// src/types/filter.type.ts
export type UserFilterType = {
    user_id: number,        // идентификатор пользователя
    position: string,       // должность
    grade: string,          // грейд
    type: string,           // тип занятости
    // salary: string,         // зп
    // location: string,       // локация
    lang: string            // яп
    channels: string[]  // каналы
}